RT86035 Fixing compile warnings

+++ arch/arm/kernel/traps.c
@@ -577,10 +577,10 @@ asmlinkage int arm_syscall(int no, struct pt_regs *regs)
 		switch (regs->ARM_r0) {
 		case SYS_ARM_ATOMIC_SET:
 		{
-			unsigned int oldval=0, flags;
+			unsigned long oldval=0, flags;
 			local_irq_save(flags);
-			oldval = *(unsigned int *)(regs->ARM_r1);
-			*(unsigned int *)(regs->ARM_r1) = regs->ARM_r2;
+			oldval = *(unsigned long *)(regs->ARM_r1);
+			*(unsigned long *)(regs->ARM_r1) = regs->ARM_r2;
 			local_irq_restore(flags);
 			return oldval;
 		}
